+++ src/idtrackerai/base/animals_detection/__init__.py
@@ -2,6 +2,7 @@ from .animals_detection import animals_detection_API
 from .segmentation import (
     generate_background_from_frame_stack,
     generate_frame_stack,
+    load_custom_background,
     process_frame,
     to_gray_scale,
 )
@@ -12,4 +13,5 @@ __all__ = [
     "generate_frame_stack",
     "process_frame",
     "to_gray_scale",
+    "load_custom_background",
 ]