+++ src/idtrackerai/segmentation_app/widgets/bkg_widget.py
@@ -9,6 +9,7 @@ from qtpy.QtWidgets import (
     QCheckBox,
     QComboBox,
     QDialog,
+    QFileDialog,
     QHBoxLayout,
     QMessageBox,
     QProgressDialog,
@@ -20,9 +21,14 @@ from qtpy.QtWidgets import (
 from idtrackerai.base.animals_detection import (
     generate_background_from_frame_stack,
     generate_frame_stack,
+    load_custom_background,
 )
-from idtrackerai.GUI_tools import Canvas
-from idtrackerai.utils import Episode
+from idtrackerai.GUI_tools import Canvas, WrappedLabel
+from idtrackerai.utils import Episode, IdtrackeraiError
+
+
+def is_standard(stat: str) -> bool:
+    return stat.lower() in ("median", "mean", "max", "min")
 
 
 class BkgComputationThread(QThread):
@@ -41,13 +47,18 @@ class BkgComputationThread(QThread):
 
     def setStat(self, stat: str):
         new_stat = stat.lower()
-        if hasattr(self, "background_stat") and new_stat == self.background_stat:
-            return
         self.background_stat = new_stat
         self.bkg = None
         if hasattr(self, "video_paths"):
             # when the App in inactive, Stat is set but there is no video_paths yet
-            self.start()
+            if not is_standard(self.background_stat):
+                # we load the custom background in the main thread because it can raise IdtrackeraiError
+                self.bkg = load_custom_background(
+                    self.background_stat, self.video_paths[0]
+                )
+                self.finished.emit()
+            else:
+                self.start()
 
     def set_parameters(
         self, video_paths: Sequence[str | Path], episodes: list[Episode]
@@ -135,13 +146,18 @@ class BkgWidget(QWidget):
         self.checkBox.stateChanged.connect(self.CheckBox_changed)
 
         self.bkg_stat = QComboBox()
-        self.bkg_stat.addItems(("Median", "Mean", "Max", "Min"))
+        self.bkg_stat.addItems(("Custom", "Median", "Mean", "Max", "Min"))
         self.bkg_stat.setCurrentIndex(-1)
         self.bkg_stat.setEnabled(False)
         self.bkg_stat.setFocusPolicy(Qt.FocusPolicy.NoFocus)
         self.bkg_stat.setSizePolicy(
             QSizePolicy.Policy.Maximum, QSizePolicy.Policy.Maximum
         )
+        self.last_stat = "Median"
+
+        self.uploaded_path = WrappedLabel(framed=True)
+        self.uploaded_path.clicked.connect(lambda: self.bkg_stat_changed("Custom"))
+        self.uploaded_path.setVisible(False)
 
         self.view_bkg = QToolButton()
         self.view_bkg.setText("View background")
@@ -155,19 +171,58 @@ class BkgWidget(QWidget):
         self.setLayout(layout)
         layout.addWidget(self.checkBox)
         layout.addWidget(self.bkg_stat)
+        layout.addWidget(self.uploaded_path)
         layout.addWidget(self.view_bkg)
-        self.bkg_stat.currentTextChanged.connect(self.bkg_thread.setStat)
+        self.bkg_stat.currentTextChanged.connect(self.bkg_stat_changed)
         self.bkg_thread.set_progress_value.connect(self.set_progress_value)
         self.bkg_thread.set_progress_max.connect(self.set_progress_maximum)
         self.bkg_thread.started.connect(self.bkg_thread_started)
         self.bkg_thread.finished.connect(self.bkg_thread_finished)
+
+    def bkg_stat_changed(self, text: str) -> None:
+        if is_standard(text):
+            self.bkg_thread.setStat(text)
+            return
+
+        assert text.lower() == "custom"
+
+        file_path, _ = QFileDialog.getOpenFileName(
+            self,
+            "Select background image",
+            self.uploaded_path.text(),
+            "Image files (*.png *.jpg *.jpeg *.bmp *.tiff);;All files (*)",
+        )
+        if not file_path:
+            self.set_bkg_stat(self.last_stat)
+            return
+
+        self.set_custom_path(file_path)
+
+    def set_custom_path(self, file_path: str):
+        try:
+            self.bkg_thread.setStat(file_path)
+        except IdtrackeraiError as err:
+            QMessageBox.critical(None, "Error", str(err))
+            self.set_bkg_stat(self.last_stat)
+            return
+
+        self.uploaded_path.setText(file_path)
+        self.bkg_stat.blockSignals(True)
+        self.bkg_stat.setCurrentText("Custom")
+        self.bkg_stat.blockSignals(False)
+
+    def set_bkg_stat(self, stat: str):
+        if stat.lower() in ("median", "mean", "max", "min", "custom"):
+            self.bkg_stat.setCurrentText(stat.capitalize())
+        else:
+            self.set_custom_path(stat)
 
     def set_new_video_paths(self, video_paths, episodes):
         self.video_paths = video_paths
         self.episodes = episodes
         self.bkg_thread.bkg = None
         self.bkg_thread.frame_stack = None
-        if not self.checkBox.isChecked():
+        if not self.checkBox.isChecked() or self.bkg_stat.currentText() == "Custom":
             return
         QMessageBox.information(
             self,
@@ -188,11 +243,16 @@ class BkgWidget(QWidget):
                 self.checkBox.setChecked(False)
                 return
             self.bkg_thread.set_parameters(self.video_paths, self.episodes)
-            self.bkg_thread.start()
+            self.bkg_thread.setStat(
+                self.bkg_stat.currentText()
+                if is_standard(self.bkg_stat.currentText())
+                else self.uploaded_path.text()
+            )
         else:
             self.view_bkg.setEnabled(False)
             self.new_bkg_data.emit(None)
         self.bkg_stat.setEnabled(checked)
+        self.uploaded_path.setEnabled(checked)
 
     def bkg_thread_started(self):
         self.progress_bar = QProgressDialog("Computing background", "Cancel", 0, 100)
@@ -208,12 +268,26 @@ class BkgWidget(QWidget):
         self.progress_bar.setMaximum(value)
 
     def bkg_thread_finished(self):
-        self.progress_bar.reset()
+        if hasattr(self, "progress_bar"):
+            self.progress_bar.reset()
         if self.bkg_thread.bkg is None:
             self.checkBox.setChecked(False)
         else:
             self.view_bkg.setEnabled(True)
+            self.last_stat = (
+                self.bkg_stat.currentText()
+                if is_standard(self.bkg_stat.currentText())
+                else self.uploaded_path.text()
+            )
+        self.uploaded_path.setVisible(not is_standard(self.last_stat))
         self.new_bkg_data.emit(self.bkg_thread.bkg)
 
     def getBkg(self):
         return self.bkg_thread.bkg if self.checkBox.isChecked() else None
+
+    def get_bkg_stat(self) -> None | str:
+        if not self.checkBox.isChecked():
+            return None
+        if self.bkg_stat.currentText() == "Custom":
+            return self.uploaded_path.text()
+        return self.bkg_stat.currentText().lower()