+++ docs/source/user_guide/changelog.rst
@@ -24,6 +24,13 @@ Changelog
 
 Authors since :ref:`5.0.0`: Jordi Torrents (jordi.torrentsm@gmail.com)
 
+6.0.10
+======
+
+- Add the possibility to upload a custom background in the Segmentation App and by using the parameter ``BACKGROUND_SUBTRACTION_STAT``, see :ref:`Background subtraction`.
+- Ask for a session to open if no session directory is provided when opening the :ref:`validator_reference`.
+
+
 6.0.8
 =====